Разделит код на отдельные модули
