+++ index.html
@@ -227,6 +227,6 @@
     </section>
   </template>
 <script src="functions.js"></script>
-<script src="main.js"></script>
+<script type="module" src="main.js"></script>
 </body>
 </html>
